Fix memory leak in vty motd and avoid wedge in vty when connection reset.
When constructing the motd message when a vty connection is made, some
temporary strings were not released.  This leaked a small amount of
memory on each connection.

If a vty connection was made and immediately reset, it could leave the
vty in an invalid state, such that the vty would loop indefinitely,
wedged at 100% CPU and doing no further useful work.

These bugs have been corrected.  Furthermore, some defensive logic has
been added to the vty state handling, so that if a vty were to become
wedged in a similar way, that would be detected and the vty closed.