Fix bug introduced into zread_ipv4_add() during merge of master.
In commit 81188036c33d098e43734da3603a37350f3272b7 on Sat 17-Mar-2012
managed to duplicate:

    message = stream_getc (s);

in zread_ipv4_add() -- which caused the reading of the message to
be out of step, and when reading the prefix could pick up an
invalid prefix length and overrun a buffer.

Duplicate line removed, and check for valid prefix length added.