Version number update: Quagga, 0.99.20ex26b
Changes since the last version include:

  * fixes for issues created when merging in 0.99.20

  * fixes to compile and run under FreeBSD

  * fixes for issues thrown up by OSRF ANVL tests

  * improved diagnostics for timers

  * updates to vtysh -- work in progress

  * detabbing some modules... see below.

Note on "detab"
===============

Some modules have been "detabbed", which is intended to make
it easier to merge the euro_ix branch with the main one.

The problem this addresses is: where an editor (eg the
Eclipse editor) automatically converts tabs to spaces and
discards trailing white-space, then when committing/merging
this creates entirely spurious changes, which clutter the
patches.

The detab process is entirely mechanical:

   * replace tabs by spaces -- assuming tab interval of 8.

   * replace all (remaining) control characters by '\n'.

     For ^L (FF) this generally introduces an (additional) blank
     line.

   * strip all trailing spaces

and is done by a little piece of perl, for all .c and .h files.

When the euro_ix stuff is merged with the main branch, if the
main branch is detabbed first, then a lot of spurious changes
can be suppressed.