Small changes to unintern and free sub-attributes.
Make all the sub attribute free functions:

  * do nothing with NULL argument -- for consistency

  * return NULL -- likewise

Make all unintern functions:

  * cope cleanly with refcnt == 0 -- see below

  * take a indirect pointer and set NULL if frees value -- see below.

  * return NULL -- unconditionally

Add attr_refcnt_t so that all refcnt can be ensured consistent.

Replace aspath_empty() and aspath_empty_get() by aspath_empty(), now
with a bool argument to return an intern/unintern empty aspath.

A refcnt of 0 <=> a new value which has not yet been interned.  The
unintern function should honour that and simply discard the value.
Handling sets of attributes with a mix of interned and uninterned
values is tricky... but this does something predictable (does not
drive the refcnt -ve !) and plausible.

Some (not all) unintern functions took an indirect pointer, they now
all do.  If a sub-attribute is freed, and the parent attribute is
not freed or discarded, then we have ourselves a BUG.  However,
clearing the pointer at least means there is no dangling reference.

If an interned attr set has a sub-attr freed when the attr has a
reference count > 1, it will not be possible to unintern the attr,
because it won't be findable.  Much worse than that, the unintern
of the broken attr may find an unbroken set.  So, for attr and for
all sub-attr, during intern, if the attr or sub-attr are not
returned from the hash, nothing is freed.  This could remove an
unbroken attr/sub-attr, which at least stay in existence, but
will not be found when they are uninterned and any new instance
will grow a new hash entry.  The effect should be to leak memory,
but stay up.  These errors are now logged.