Tighten up handling of attribute sets, and temporary attribute sets
This does not actually change a great deal, but establishes the ground
rules for manipulating sets of attributes.

The problem is how to handle changes to a set of attributes, for example
changes implemented by a route-map.

The mechanism settled on is:

  * start with a set of attributes where all (existing) sub-attr are
    interned.

NB: the attr set itself does NOT need to be interned.

  * do bgp_attr_dup().  This does not change the state of the sub-attr,
    but creates a temporary copy of the given attr, where the
    temporary has refcnt == 0 (whatever given attr has).

  * make changes to the temporary attr:

  * simply replacing any interned sub-attr by new, uninterned
    values.

  * updating any uninterned values.  (Freeing as required.)

  * either:

  * interning the result -- bgp_attr_intern_temp()

  * discarding the result -- bgp_attr_flush()

AND then forgetting the temporary attributes.

The logic is that the original "owns" a refcnt on the original interned
sub-attr (so must not be uninterned until after the temporary
attributes have beed either interned or flushed).  While the temporary
has no refcnts of its own, but does "own" any uninterned sub-attr.

This is a little restrictive, but safe.

It is a rule that all sub-attr MUST have a refcnt >= the refcnt of
any give attr set (interned or not).

It is possible to unintern an attr which has refcnt == 0.  That will
unintern all sub-attr and free any whose refcnt is <= 1.  This may be
It is a rule that all sub-attr MUST have a refcnt >= the refcnt of
any give attr set (interned or not).

It is possible to unintern an attr which has refcnt == 0.  That will
unintern all sub-attr and free any whose refcnt is <= 1.  This may be
used when a attr set is constructed from scratch with interned
sub-attr (only), and is now to be dismantled -- see bgp_update_receive()

bgp_attr_unintern() used to take an indirect pointer, and clear that
if the attr was freed.  That was of little use.  Generally, after
uninterning the attr the caller has lost interest in the attr
altogether, so does not care whether it was removed from the hash
or not !

bgp_advertise_unintern () is brought into line with other unintern.