Tidy up attribute handling and route-map invocation in bgp_route.
This should work as before.

Parsed attributes have all sub-attributes interned, but are passed in
as an uninterned set.  At other times, fully interned sets are
handled -- eg at soft-reconfig.  It doesn't matter whether the
attr themselves are interned -- so that is fine.  Here we are careful
to ensure that we are handling attrs with interned sub-attrs, and
tread carefully through route-maps.

The peer->rmap_type handling was a bit magic... depending on it being
cleared after use and kept clear.  Have simplified that.