Return interned attr set from bgp_announce_check()
Also bgp_announce_check_rsclient().

This is a lot tidier than handling a temporary set of attributes
across a number of functions.

The existing code appears to work, so this should have minimal
impact.

It is essential that announcements for RS-Clients go into the
RS-CLient RIB, and not the Main RIB -- so added some qdebug checks
for that.

Return from bgp_process_announce_selected() was not required,
so changed that to void function.