Clean up how attributes are constructed for default/static/redistribute.
It is simpler to construct a temporary set of attributes with *nothing*
interned, then run that through the route-map stuff and intern the
result.

This saves some dicking about with interned aspath stuff.

For redistribute, having made a basic set of attributes, those are
interned, and then run through the route-map for each bgp instance.
So that changes made by the route-map can be identified and dealt
with correctly.

No change in behaviour, except for redistribution, where would be
different if there were more than one instance enabled for
redistribution -- which there isn't.