Review and fix sub-attr chages in route-map 'set' functions.
When applying a route-map, the rule is that:

  * all interned values "belong" to some other set of attr,
    so the can simply be replaced -- by uninterned values.

  * all uninterned values "belong" to this set of attr, so
    the 'set' function is responsible for freeing any
    uninterned values it replaces.

set as-path prepend was OK, but did not cope with the (screamingly
unlikely) case of a NULL as-path.

set as-path exclude likewise.

set community was almost OK, but leaked memory in the "none" case and
when replacing a community iff the community had already been changed
(and hence was uninterned).

set comm-list xxx delete was OK.

set extcommunity rt and set extcommunity soo were both BROKEN.  They
uninterned any existing value, which was not theirs to unintern, and
interned a new value, also wrong.