Review and tighten adj-out and advertisment handling.
The bgp_adj_out and bgp_advertise objects are closely bound together,
so that changes to the adj-out immediately start/stop/restart
advertisements on the relevant queue for the peer.  When an
advertisement is executed -- so an UPDATE message is sent to the peer,
the adj-out is updated (and after a withdraw, removed).  All the
pieces of information needed by the advertisement are locked by it,
so this is all nicely self contained, EXCEPT for the bgp_info
structure for update advertisements.

When a route is withdrawn its related bgp_info structure lives on
until it is given to bgp_info_reap() by the bgp_node process
function.  As far as one can see, all bgp_advertise objects
which refer to the reaped bgp_info will be stopped or turned into
withdraws -- and so drop their pointer to the bgp_info structure.

Just in case that does not happen, bgp_info_reap() now sets the
bgp_info to BGP_INFO_REAPED.  And if an advertisement is then
executed for a bgp_advertise object which points at such a bgp_info
it will be handled as a withdraw -- avoiding doing anything with
information which is definitely out of date !

There is no evidence that this ever happens.  But if it does, it is
now coped with AND logged as an error.