Change how bgp_announce_check() deals with Route Reflection & route-map.
When refelecting, the 'out' route-map is not allowed to change the
attributes.

This was being done by copying the temporary attributes to a second
temporary set, and then discarding that set.  That will fail if the
aspath has just been set to an empty path and the Route-Map then sets
something in the aspath -- because that will free the recently created
empty path, leaving a dangling pointer in the attribute set which is
about to be interned.  It would fail similarly if any other sub-attr
were uninterned.

So, the procedure now interns the attributes to date before applying
the route-map -- iff the route is being reflected.  If the route is
not denied, this is simply doing the necessary intern operation a step
or two earlier.