Improve announcement logic in bgp_process_rsclient()
Generally, if the selection has not changed and the attributes for the
current selection are not BGP_INFO_ATTR_CHANGED, the processing logic
will avoid generating an announcement.

This was not the case for ordinary Route-Server Clients, so a
Route-Server would send the same route more than once.  This is
corrected by this change.

For Groups of Route Server Clients, it did avoid sending duplicate
updates, but would check for the same thing for each group
member !  (OK, but a waste of time).

Further, for Groups of Route Server clients, made the unwarranted
assumption that all group members are members for all afi/safi,
which was seriously broken.