Improve announce logic in bgp_process_main().
Having decided on a new selection in the Main RIB,
bgp_process_main() announces the result to all peers, including all
RS-Clients.

This has all the appearance of a huge mistake.  However, it is probably
just a sad waste of effort.

bgp_process_announce_selected() runs bgp_announce_check() or
bgp_announce_check_rsclient(), depending in the table type for the
current bgp_node.  In this case, that is BGP_TABLE_MAIN, so the
route will be given to bgp_announce_check(), which will reject
it because it will not announce stuff from the Main RIB to any
RS-Client.  The result of that is to bgp_adj_out_unset()... which
will have nothing to do, because the RS-Client does not have
adj-out entries in the Main RIB.

So, for RS-Clients this is a lengthy-ish NO-OP !

Added a check to avoid calling bgp_process_announce_selected() in
bgp_process_main() for RS-Clients.