Review use of bgp_clear_rsclient_rib()
Changes made for greater clarity.  Function unaffected.

Name notwithstanding, this is not really a "clear" operation.

This is used when an RS-Client RIB has been emptied of all routes,
and the entire RIB is due to be freed.

Changed name to bgp_finish_rsclient_rib() and included the
bgp_table_finish() which always followed bgp_clear_rsclient_rib().

This function removes all routes contributed to the RS-Client RIB
by all peers.  There is a cosmetic change here:

 -  bgp_rib_remove (rn, ri, rsclient, table->afi, table->safi);
 +  bgp_rib_remove (rn, ri, ri->peer, table->afi, table->safi);

...bgp_rib_remove() is generally called with the peer which owns
the bgp_info (ri) -- the source peer.  The old code was calling
it with the peer which owns the RIB -- the destination peer.  As
it happens this makes no difference, the peer is only used to find
the bgp instance -- which is the same for both.

The function performs a belt-and-braces clear out of all adj-out
and adj-in.  All RS-Clients must have been cleared by this time,
so there should be no adj-out.  There should never be any adj-in.

In bgp_peer_delete() three loops dealing with RS-Client RIB have
been collapsed into one, without changing the effect but (hopefully)
clearer as to purpose.