Review bgp_clear_routes() and bgp_clear_adj_in()
bgp_clear_routes() clears all routes (a) contributed to all RIBs
by the given peer, and (b) announced to the given peer.  Added
some comments, but seems OK.

bgp_clear_adj_in() removes the remembered state of routes as
received from the given peer.  This information lives in the
Main RIB.  Previous code walked the main RIB looking for adj-in
entries.  New code uses the list of known adj-in for the peer
(as used elsewhere for clearing the peer).

Behaviour should be unaffected.