Fix MPLS and RS-Client and Soft-Reconfig.
For soft-reconfig the adj-in now carries a copy of the tag as
received from the peer.

[Previous code seems to assume that there will be a bgp_info,
and that its tag should be used.  It's not clear that there
will be a bgp_info -- since the route in the adj-in may have
been filtered out !  It's not clear that there will be only
one bgp_info -- so it's not clear who's tag this is using.
The previous behaviour appears broken.]

Soft-reconfig operations now use the tag in the adj-in.

RS-Client soft-reconfig will now work for MPLS RS !

The bgp_finish_rsclient_rib() will now work for MPLS RS !

[It is not clear that for non-RS MPLS RIBs that those are
correctly dismantled.  There does not seem to be a mechanism
to clear out redundant RD entries, or a mechanism to walk two
levels of table to dismantle same.]

bgp_update_main()/_rsclient() will now detect MPLS tag changes
as well as attribute changes.  Previous code would not notice
a tag change if the attributes did not also change !