Review bgp_announce_route()
This did something peculiar.  For RS-Clients it processed the Main RIB
and attempted to announce the current selection for all prefixes.
This would have been a disaster, except that it used
bgp_announce_check() -- the correct check for the Main RIB -- which
denies all MAIN RIB routes to RS-Clients.  So the result is just to
waste a certain amount of time.  Afterwards, for RS-Clients it would
then announce the RS-Client RIB.

The code now announces the Main RIB to Main RIB peers, and the RS-RIB
to RS-Clients.

For MPLS RS-Clients the code was broken, and has now been corrected.