Declutter bgp_afi_node_get() and add bgp_afi_node_lookup()
There is no need to pass afi/safi to bgp_afi_node_get(), since it is
passed a table which implies that.

The setting of rn->prn is required only when a new node is created.

A Route Distinguisher sub-table should have the same type as its
parent.  This is the only behaviour change.

The logic for creating and walking MPLS sub-tables should not be
repeated.