Avoid memory-leak in lib/thread.c when running pThreaded
When running pThreaded the lib/thread.c THREAD_TIMER object uses
a lib/qtimer timer, and must release that when the object is
dequeued from that queue.  This was mostly OK, but is now
tighter by making that the responsibility of the remove from
queue function -- where it cannot be missed.

Also: fixed the distinction between the queue an object is on
and the type of the thread.  This improves the diagnostics
for types of thread and CPU utilization.